// packages/visualiser/src/replay/generated/replay.ts
export const Team = {
    A: 0,
    B: 1,
} as const;
export type Team = (typeof Team)[keyof typeof Team];
export const Direction = {
    NORTH: 0,
    EAST: 1,
    SOUTH: 2,
    WEST: 3,
} as const;
export type Direction = (typeof Direction)[keyof typeof Direction];
export const DragonDeathReason = {
    HIT_WALL: 0,
    HIT_SELF: 1,
    HIT_OTHER_BODY: 2,
    HIT_HEAD_TO_HEAD: 3,
    NO_VALID_ACTION: 4,
} as const;
export type DragonDeathReason = (typeof DragonDeathReason)[keyof typeof DragonDeathReason];
export const DebugShape = {
    LINE: 0,
    DOT: 1,
} as const;
export type DebugShape = (typeof DebugShape)[keyof typeof DebugShape];
export const GameEndReason = {
    TEAM_ELIMINATED: 0,
    ROUND_LIMIT: 1,
} as const;
export type GameEndReason = (typeof GameEndReason)[keyof typeof GameEndReason];
export class Point extends $.Struct {
    static readonly _capnp = {
        displayName: "Point",
        id: "f46b7269eb2bbb4c",
        size: new $.ObjectSize(8, 0),
    };
    get x(): number {
        return $.utils.getInt32(0, this);
    }
    set x(value: number) {
        $.utils.setInt32(0, value, this);
    }
    get y(): number {
        return $.utils.getInt32(4, this);
    }
    set y(value: number) {
        $.utils.setInt32(4, value, this);
    }
    toString(): string {
        return "Point_" + super.toString();
    }
}
export class DebugDraw extends $.Struct {
    static readonly _capnp = {
        displayName: "DebugDraw",
        id: "eb39861b647f4934",
        size: new $.ObjectSize(8, 2),
    };
    get shape(): DebugShape {
        return $.utils.getUint16(0, this) as DebugShape;
    }
    set shape(value: DebugShape) {
        $.utils.setUint16(0, value, this);
    }
    _adoptFrom(value: $.Orphan<Point>): void {
        $.utils.adopt(value, $.utils.getPointer(0, this));
    }
    _disownFrom(): $.Orphan<Point> {
        return $.utils.disown(this.from);
    }
    get from(): Point {
        return $.utils.getStruct(0, Point, this);
    }
    _hasFrom(): boolean {
        return !$.utils.isNull($.utils.getPointer(0, this));
    }
    _initFrom(): Point {
        return $.utils.initStructAt(0, Point, this);
    }
    set from(value: Point) {
        $.utils.copyFrom(value, $.utils.getPointer(0, this));
    }
    _adoptTo(value: $.Orphan<Point>): void {
        $.utils.adopt(value, $.utils.getPointer(1, this));
    }
    _disownTo(): $.Orphan<Point> {
        return $.utils.disown(this.to);
    }
    get to(): Point {
        return $.utils.getStruct(1, Point, this);
    }
    _hasTo(): boolean {
        return !$.utils.isNull($.utils.getPointer(1, this));
    }
    _initTo(): Point {
        return $.utils.initStructAt(1, Point, this);
    }
    set to(value: Point) {
        $.utils.copyFrom(value, $.utils.getPointer(1, this));
    }
    get red(): number {
        return $.utils.getUint8(2, this);
    }
    set red(value: number) {
        $.utils.setUint8(2, value, this);
    }
    get green(): number {
        return $.utils.getUint8(3, this);
    }
    set green(value: number) {
        $.utils.setUint8(3, value, this);
    }
    get blue(): number {
        return $.utils.getUint8(4, this);
    }
    set blue(value: number) {
        $.utils.setUint8(4, value, this);
    }
    toString(): string {
        return "DebugDraw_" + super.toString();
    }
}
export const PlayerAction_Which = {
    MOVE: 0,
    SPLIT: 1,
    SUICIDE: 2,
} as const;
export type PlayerAction_Which = (typeof PlayerAction_Which)[keyof typeof PlayerAction_Which];
export class PlayerAction extends $.Struct {
    static readonly MOVE = PlayerAction_Which.MOVE;
    static readonly SPLIT = PlayerAction_Which.SPLIT;
    static readonly SUICIDE = PlayerAction_Which.SUICIDE;
    static readonly _capnp = {
        displayName: "PlayerAction",
        id: "b5bb6d59b34e5fde",
        size: new $.ObjectSize(8, 1),
    };
    _adoptMove(value: $.Orphan<$.List<Direction>>): void {
        $.utils.setUint16(0, 0, this);
        $.utils.adopt(value, $.utils.getPointer(0, this));
    }
    _disownMove(): $.Orphan<$.List<Direction>> {
        return $.utils.disown(this.move);
    }
    get move(): $.List<Direction> {
        $.utils.testWhich("move", $.utils.getUint16(0, this), 0, this);
        return $.utils.getList(0, $.Uint16List, this) as $.List<Direction>;
    }
    _hasMove(): boolean {
        return !$.utils.isNull($.utils.getPointer(0, this));
    }
    _initMove(length: number): $.List<Direction> {
        $.utils.setUint16(0, 0, this);
        return $.utils.initList(0, $.Uint16List, length, this) as $.List<Direction>;
    }
    get _isMove(): boolean {
        return $.utils.getUint16(0, this) === 0;
    }
    set move(value: $.List<Direction>) {
        $.utils.setUint16(0, 0, this);
        $.utils.copyFrom(value, $.utils.getPointer(0, this));
    }
    get split(): number {
        $.utils.testWhich("split", $.utils.getUint16(0, this), 1, this);
        return $.utils.getInt32(4, this);
    }
    get _isSplit(): boolean {
        return $.utils.getUint16(0, this) === 1;
    }
    set split(value: number) {
        $.utils.setUint16(0, 1, this);
        $.utils.setInt32(4, value, this);
    }
    get _isSuicide(): boolean {
        return $.utils.getUint16(0, this) === 2;
    }
    set suicide(_: true) {
        $.utils.setUint16(0, 2, this);
    }
    toString(): string {
        return "PlayerAction_" + super.toString();
    }
    which(): PlayerAction_Which {
        return $.utils.getUint16(0, this) as PlayerAction_Which;
    }
}
// ...
export const SonarHitKind = {
    UNKNOWN: 0,
    EMPTY: 1,
    KELP: 2,
    ALLY: 3,
    ALLY_HEAD: 4,
    ENEMY: 5,
    ENEMY_HEAD: 6,
} as const;
export type SonarHitKind = (typeof SonarHitKind)[keyof typeof SonarHitKind];